Reestructurando hook de dibujado 💀
